Refactoring: Abstract notification history service using interface
