+++ src/main/java/com/honeypot/domain/notification/api/NotificationApi.java
@@ -2,15 +2,12 @@ package com.honeypot.domain.notification.api;
 
 import com.honeypot.common.model.exceptions.InvalidTokenException;
 import com.honeypot.common.utils.SecurityUtils;
-import com.honeypot.common.validation.constraints.AllowedSortProperties;
 import com.honeypot.domain.notification.dto.NotificationResource;
 import com.honeypot.domain.notification.dto.NotificationTokenDto;
-import com.honeypot.domain.notification.repository.NotificationRepository;
 import com.honeypot.domain.notification.service.NotificationHistoryService;
 import com.honeypot.domain.notification.service.NotificationTokenManageService;
 import com.honeypot.domain.notification.dto.NotificationTokenUploadRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
